requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.26.160
botocore==1.29.160
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.3
curlify==2.2.1
daphne==4.0.0
Django==3.2.20
django-cors-headers==4.1.0
django-environ==0.10.0
django-picklefield==3.1
django-q==1.3.9
django-sesame==3.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-tracking==1.8.4
exceptiongroup==1.1.3
facebook-business==18.0.2
frozenlist==1.3.3
google-ads==21.2.0
google-api-core==2.11.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
linkedin-api-client==0.3.0
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.7.0.72
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
